merge

+++ src/main/java/io/swyp/luckybackend/users/controller/UserController.java
@@ -1,11 +1,11 @@
 package io.swyp.luckybackend.users.controller;
 
-import io.swagger.v3.oas.annotations.media.Content;
-import io.swagger.v3.oas.annotations.media.Schema;
 import io.swyp.luckybackend.common.ResponseDTO;
+import io.swyp.luckybackend.users.domain.UserEntity;
 import io.swyp.luckybackend.users.dto.ModifyUserRequestDto;
 import io.swyp.luckybackend.users.service.KakaoService;
 import io.swyp.luckybackend.users.service.UserService;
+import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -15,7 +15,6 @@ import org.springframework.web.bind.annotation.*;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 
-import java.net.http.HttpHeaders;
 
 @RestController
 @RequestMapping("/users")
@@ -26,29 +25,35 @@ public class UserController {
     private final UserService userServiceImpl;
     private final KakaoService kakaoService;
 
+
     @Operation(
             summary = "로그인 API"
     )
     @GetMapping("/sign-in")
     public void login(HttpServletResponse response) throws Exception {
         log.info("로그인 API 진입");
-        response.sendRedirect("/oauth2/authorization/kakao");
+        response.sendRedirect("/lucky/oauth2/authorization/kakao");
     }
 
     @Operation(
             summary = "회원 정보 조회 API"
     )
     @GetMapping("")
-    public ResponseEntity<ResponseDTO> getUserInfo(@RequestHeader("Authorization") String token) throws Exception {
+    public ResponseEntity<ResponseDTO> getUserInfo(HttpServletRequest request) throws Exception {
         log.info("회원 정보 조회 API 진입");
+        String token = request.getHeader("Authorization");
         return userServiceImpl.getUserInfo(token);
 
     }
 
     @Operation(summary = "회원 정보 수정 API")
     @PutMapping("")
-    public ResponseEntity<ResponseDTO> modifyUserInfo(@RequestHeader("Authorization") String token, @RequestBody ModifyUserRequestDto requestDto) throws Exception {
+    public ResponseEntity<ResponseDTO> modifyUserInfo(HttpServletRequest request, @RequestBody ModifyUserRequestDto requestDto) throws Exception {
         log.info("회원 정보 수정 API 진입");
+        if (requestDto.getNickname().length() > 9) {
+            return ResponseDTO.exceedNicknameLength();
+        }
+        String token = request.getHeader("Authorization");
         userServiceImpl.modifyUserInfo(token, requestDto.modifyDto2Entity(requestDto.getNickname(), requestDto.getEmail()));
         return ResponseDTO.success("ok");
     }
@@ -71,20 +76,16 @@ public class UserController {
 
     @Value("${kakao-admin-key}")
     private String adminKey;
+
     @Operation(
             summary = "회원 탈퇴 API"
     )
     @DeleteMapping("")
-    public ResponseEntity<ResponseDTO> withdrawUser(@RequestHeader("Authorization") String token) throws Exception {
+    public ResponseEntity<ResponseDTO> withdrawUser(HttpServletRequest request) throws Exception {
         log.info("회원 탈퇴 API 진입");
-        long userNo = userServiceImpl.deleteUser(token);
-        return ResponseDTO.success(kakaoService.unlinkUser(userNo, adminKey).block());
-    }
-
-
-    @DeleteMapping("/test")
-    public ResponseEntity<ResponseDTO> test(@RequestHeader("Authorization") String token) throws Exception {
-        log.info("TEST API 진입");
+        String token = request.getHeader("Authorization");
+        boolean isExist = userServiceImpl.isExistUser(token);
+        if (!isExist) return ResponseDTO.notExistedUser();
         long userNo = userServiceImpl.deleteUser(token);
         return ResponseDTO.success(kakaoService.unlinkUser(userNo, adminKey).block());
     }
